Add radar_emulator command that allows reconfiguring the internal trigger on the fly

+++ include/timer/emulator.h
@@ -1,2 +1,2 @@
 
-void radar_emulator_start(uint32_t target_hz, double target_t);
+void radar_emulator_start(double target_hz, double target_t);
